# swagger.yaml
openapi: 3.0.3
info:
  title: Reach Leads API
  version: 1.0.0
  description: API documentation for the Reach Leads Node.js and Express backend.
servers:
  - url: http://localhost:5000
    description: Local
  - url: https://your-app-name.onrender.com
    description: Production
...
paths:
...
  /api/leads:
    get:
      tags:
        - Leads
      summary: Get all leads
      operationId: getLeads
      responses:
        "200":
          description: Leads returned successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessListResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Lead"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - Leads
      summary: Create a lead
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/leads/{id}:
    get:
      tags:
        - Leads
      summary: Get a lead
      operationId: getLead
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Lead returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Leads
      summary: Update a lead
      operationId: updateLead
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdateInput"
      responses:
        "200":
          description: Lead updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Leads
      summary: Delete a lead
      operationId: deleteLead
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Lead deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/campaigns:
    get:
      tags:
        - Campaigns
      summary: Get all campaigns
      operationId: getCampaigns
      responses:
        "200":
          description: Campaigns returned successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessListResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Campaign"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - Campaigns
      summary: Create a campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/campaigns/{id}:
    get:
      tags:
        - Campaigns
      summary: Get a campaign
      operationId: getCampaign
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Campaign returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Campaigns
      summary: Update a campaign
      operationId: updateCampaign
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdateInput"
      responses:
        "200":
          description: Campaign updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Campaigns
      summary: Delete a campaign
      operationId: deleteCampaign
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Campaign deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/outreach/send:
    post:
      tags:
        - Outreach
      summary: Create outreach message
      operationId: createOutreach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachInput"
      responses:
        "201":
          description: Outreach message created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/outreach/history:
    get:
      tags:
        - Outreach
      summary: Get outreach history
      operationId: getOutreachHistory
      responses:
        "200":
          description: Outreach history returned successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessListResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Outreach"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/accounts:
    get:
      tags:
        - Accounts
      summary: Get connected accounts
      operationId: getAccounts
      responses:
        "200":
          description: Accounts returned successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessListResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/SocialAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - Accounts
      summary: Connect an account
      operationId: connectAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialAccountInput"
      responses:
        "201":
          description: Account connected successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Get an account
      operationId: getAccount
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Account returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Accounts
      summary: Delete an account
      operationId: deleteAccount
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Account removed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/accounts/{id}/disconnect:
    patch:
      tags:
        - Accounts
      summary: Disconnect an account
      operationId: disconnectAccount
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Account disconnected successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/ai/generate:
    post:
      tags:
        - AI
      summary: Generate AI content
      operationId: generateContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerateRequest"
      responses:
        "200":
          description: AI content generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiGenerationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/ai/rewrite:
    post:
      tags:
        - AI
      summary: Rewrite a message
      operationId: rewriteMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRewriteRequest"
      responses:
        "200":
          description: Message rewritten successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: MongoDB ObjectId.
      schema:
        type: string
        example: 64b7f2f62a37d55c9f1c1234
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ObjectId:
      type: string
      example: 64b7f2f62a37d55c9f1c1234
    TimestampFields:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    SuccessListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            count:
              type: integer
              example: 1
    MessageResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            message:
              type: string
              example: Resource deleted
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Invalid credentials
...
    Platform:
      type: string
      enum:
        - linkedin
        - twitter
        - instagram
        - facebook
        - email
    LeadSource:
      type: string
      enum:
        - linkedin
        - twitter
        - instagram
        - facebook
        - website
        - manual
    LeadStatus:
      type: string
      enum:
        - new
        - contacted
        - qualified
        - proposal
        - won
        - lost
    LeadInput:
      type: object
      required:
        - firstName
        - email
      properties:
        firstName:
          type: string
          example: Alex
        lastName:
          type: string
          example: Morgan
        email:
          type: string
          format: email
          example: alex@example.com
        phone:
          type: string
          example: "+15551234567"
        company:
          type: string
          example: Example Co
        jobTitle:
          type: string
          example: Head of Sales
        source:
          $ref: "#/components/schemas/LeadSource"
        status:
          $ref: "#/components/schemas/LeadStatus"
        notes:
          type: string
          example: Interested in outbound automation.
        campaign:
          $ref: "#/components/schemas/ObjectId"
    LeadUpdateInput:
      allOf:
        - $ref: "#/components/schemas/LeadInput"
    Lead:
      allOf:
        - type: object
          properties:
            _id:
              $ref: "#/components/schemas/ObjectId"
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
            company:
              type: string
            jobTitle:
              type: string
            source:
              $ref: "#/components/schemas/LeadSource"
            status:
              $ref: "#/components/schemas/LeadStatus"
            notes:
              type: string
            campaign:
              oneOf:
                - $ref: "#/components/schemas/ObjectId"
                - type: object
                  properties:
                    _id:
                      $ref: "#/components/schemas/ObjectId"
                    name:
                      type: string
                    platform:
                      $ref: "#/components/schemas/Platform"
            owner:
              $ref: "#/components/schemas/ObjectId"
        - $ref: "#/components/schemas/TimestampFields"
    LeadResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Lead"
    CampaignStatus:
      type: string
      enum:
        - draft
        - active
        - paused
        - completed
    CampaignInput:
      type: object
      required:
        - name
        - platform
      properties:
        name:
          type: string
          example: Q3 LinkedIn Leads
        description:
          type: string
          example: Campaign targeting SaaS founders.
        platform:
          $ref: "#/components/schemas/Platform"
        status:
          $ref: "#/components/schemas/CampaignStatus"
        budget:
          type: number
          example: 500
        targetLeads:
          type: number
          example: 100
        generatedLeads:
          type: number
          example: 0
        messagesSent:
          type: number
          example: 0
        responseRate:
          type: number
          example: 0
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    CampaignUpdateInput:
      allOf:
        - $ref: "#/components/schemas/CampaignInput"
    Campaign:
      allOf:
        - type: object
          properties:
            _id:
              $ref: "#/components/schemas/ObjectId"
            name:
              type: string
            description:
              type: string
            platform:
              $ref: "#/components/schemas/Platform"
            status:
              $ref: "#/components/schemas/CampaignStatus"
            budget:
              type: number
            targetLeads:
              type: number
            generatedLeads:
              type: number
            messagesSent:
              type: number
            responseRate:
              type: number
            startDate:
              type: string
              format: date-time
            endDate:
              type: string
              format: date-time
            owner:
              $ref: "#/components/schemas/ObjectId"
        - $ref: "#/components/schemas/TimestampFields"
    CampaignResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Campaign"
...
    OutreachStatus:
      type: string
      enum:
        - draft
        - scheduled
        - sent
        - delivered
        - opened
        - replied
        - failed
    OutreachInput:
      type: object
      required:
        - lead
        - campaign
        - platform
        - message
      properties:
        lead:
          $ref: "#/components/schemas/ObjectId"
        campaign:
          $ref: "#/components/schemas/ObjectId"
        platform:
          $ref: "#/components/schemas/Platform"
        subject:
          type: string
          example: Quick idea for your team
        message:
          type: string
          example: Hi Alex, I noticed your team is hiring SDRs...
        status:
          $ref: "#/components/schemas/OutreachStatus"
        scheduledFor:
          type: string
          format: date-time
...
    Outreach:
      allOf:
        - type: object
          properties:
            _id:
              $ref: "#/components/schemas/ObjectId"
            lead:
              oneOf:
                - $ref: "#/components/schemas/ObjectId"
                - type: object
                  properties:
                    _id:
                      $ref: "#/components/schemas/ObjectId"
                    firstName:
                      type: string
                    lastName:
                      type: string
                    email:
                      type: string
                      format: email
            campaign:
              oneOf:
                - $ref: "#/components/schemas/ObjectId"
                - type: object
                  properties:
                    _id:
                      $ref: "#/components/schemas/ObjectId"
                    name:
                      type: string
            sender:
              $ref: "#/components/schemas/ObjectId"
            platform:
              $ref: "#/components/schemas/Platform"
            subject:
              type: string
            message:
              type: string
            status:
              $ref: "#/components/schemas/OutreachStatus"
            scheduledFor:
              type: string
              format: date-time
            sentAt:
              type: string
              format: date-time
            repliedAt:
              type: string
              format: date-time
        - $ref: "#/components/schemas/TimestampFields"
    OutreachResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Outreach"
    SocialAccountStatus:
      type: string
      enum:
        - connected
        - disconnected
        - expired
    SocialAccountInput:
      type: object
      required:
        - platform
        - accountName
        - accountId
      properties:
        platform:
          $ref: "#/components/schemas/Platform"
        accountName:
          type: string
          example: Jane Doe LinkedIn
        accountId:
          type: string
          example: linkedin-123
        accessToken:
          type: string
          nullable: true
        refreshToken:
          type: string
          nullable: true
        tokenExpiresAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/SocialAccountStatus"
        lastSyncedAt:
          type: string
          format: date-time
    SocialAccount:
      allOf:
        - type: object
          properties:
            _id:
              $ref: "#/components/schemas/ObjectId"
            owner:
              $ref: "#/components/schemas/ObjectId"
            platform:
              $ref: "#/components/schemas/Platform"
            accountName:
              type: string
            accountId:
              type: string
            accessToken:
              type: string
              nullable: true
            refreshToken:
              type: string
              nullable: true
            tokenExpiresAt:
              type: string
              format: date-time
            status:
              $ref: "#/components/schemas/SocialAccountStatus"
            lastSyncedAt:
              type: string
              format: date-time
        - $ref: "#/components/schemas/TimestampFields"
    SocialAccountResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SocialAccount"
...
    AiGenerationType:
      type: string
      enum:
        - outreach
        - followup
        - campaign
        - rewrite
        - subject
    AiGenerateRequest:
      type: object
      required:
        - prompt
        - type
      properties:
        prompt:
          type: string
          example: Write a short campaign intro for SaaS founders.
        type:
          $ref: "#/components/schemas/AiGenerationType"
...
    AiRewriteRequest:
      type: object
      properties:
        message:
          type: string
          example: Can we schedule a call?
        tone:
          type: string
          example: friendly
...
    AiGeneration:
      allOf:
        - type: object
          properties:
            _id:
              $ref: "#/components/schemas/ObjectId"
            user:
              $ref: "#/components/schemas/ObjectId"
            type:
              $ref: "#/components/schemas/AiGenerationType"
            prompt:
              type: string
            output:
              type: string
            model:
              type: string
              example: gpt-4o-mini
        - $ref: "#/components/schemas/TimestampFields"
    AiGenerationResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AiGeneration"
    AiMessageResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            message:
              type: string
              example: Hi Alex, I noticed your team is growing...
...
security:
  - bearerAuth: []